populates new seqfeature_dbxref [you must have post-capetown branch of biosql-schema to have this table]
other postgres fixes - uses insert_nopk to get rid of annoying
warning when trying get_last_id() for tables with no primary key

added some transactional awareness

svn path=/bioperl-db/branches/post-capetown/; revision=335

+++ Bio/DB/SQL/SeqLocationAdaptor.pm
@@ -357,15 +357,18 @@ sub remove_by_dbID{
    	my ($self) = shift;
 	
 	my ($dbID) = join (",",@_); 
-	
+	return unless $dbID;
         my $loc_ids =
           join(", ",
                $self->select_colvals("seqfeature_location",
                                      "seqfeature_id in ($dbID)",
                                      "seqfeature_location_id")
               );
-	my $sth = $self->prepare("DELETE FROM location_qualifier_value WHERE seqfeature_location_id IN($loc_ids)");
-	$sth->execute();
+        my $sth;
+        if ($loc_ids) {
+            $sth = $self->prepare("DELETE FROM location_qualifier_value WHERE seqfeature_location_id IN($loc_ids)");
+            $sth->execute();
+        }
 	$sth = $self->prepare("DELETE FROM seqfeature_location WHERE seqfeature_id IN($dbID)");
 	$sth->execute();
